requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
algomojo==1.4
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.1
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
crypto==1.4.1
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.1.5
djangorestframework==3.12.2
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.3.1
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.3
jupyter-server-terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
Naked==0.1.32
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
NorenRestApiPy @ file:///C:/Users/TradeGui/TradeGui/dist/NorenRestApiPy-0.0.20-py2.py3-none-any.whl
notebook==7.0.3
notebook-shim==0.2.3
numpy==1.24.2
overrides==7.4.0
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==3.10.0
playsound==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.9.9
pycryptodomex==3.9.9
pydub==0.25.1
pygame==2.5.0
Pygments==2.16.1
PyJWT==2.7.0
PyMySQL==1.0.2
pynput==1.7.6
pyotp==2.8.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2020.5
pywin32==306
pywinpty==2.0.11
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
Send2Trash==1.8.2
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spark-parser==1.8.9
sqlparse==0.4.1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
twilio==8.3.0
typing-extensions==4.7.1
tzdata==2023.3
uncompyle6==3.9.0
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.8
xdis==6.0.5
yarl==1.9.2
zipp==3.16.2
